requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0
aiomysql>=0.2.0

# 向量数据库
pymilvus>=2.5.0

# HTTP客户端
httpx>=0.25.0
aiohttp>=3.9.0

# 配置管理
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 日志
loguru>=0.7.0

# 图像处理
Pillow>=10.1.0
imagehash>=4.3.1

# 文本处理
jieba>=0.42.1

# 阿里云SDK
dashscope>=1.14.0

# 工具库
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.0
openpyxl>=3.1.0  # 读取 xlsx